requirements.txt
fastapi
uvicorn
numpy
transformers
torch==2.9.0+cu128
torchaudio==2.9.0+cu128
torchcodec==0.8.1
pyannote.audio

--extra-index-url https://download.pytorch.org/whl/cu128
docker 